requirements.txt
PyQt5
Babel
black
numpy
requests
opencv-python
pysqlcipher3
sqlcipher3
python-dotenv
nuitka
pyinstaller